Moving to Components and Weapon ( now Laser is a Weapon, not a Entity )
