--- src/main/components/interfaces/Collidable.java
@@ -1,5 +1,0 @@
-package main.components.interfaces;
-
-public interface Collidable {
-    boolean wasHit(Entity entityOne, Entity entityTwo);
-}
