+++ src/main/models/components/interfaces/Collidable.java
@@ -1,0 +1,5 @@
+package main.models.components.interfaces;
+
+public interface Collidable {
+    boolean wasHit(Entity entity, Weapon deployedWeapon);
+}
