+++ src/main/models/components/interfaces/Weapon.java
@@ -1,0 +1,5 @@
+package main.models.components.interfaces;
+
+public interface Weapon {
+    void deploy();
+}
