+++ src/main/utils/WeaponType.java
@@ -1,0 +1,6 @@
+package main.utils;
+
+public enum WeaponType {
+    NPC,
+    PLAYER
+}
